UPBGE: Implement resource collection.
Previously to access to a resource (mesh, action) or getting a converted
object by name, both place where used : BL_Converter and SCA_LogicManager.

To avoid managing two classes BL_ResourceCollection is introduced. Its
goal is to own all the material, mesh and action resource and track a name
map to mesh, action and game object from conversion stage.

This class is owned by KX_Scene and instanciated by moving content of a
BL_SceneConverter (that is still needed because of the map to blender object
is maintain during conversion).