UPBGE: Simplify RAS_Rasterizer::SetViewMatrix.
Previously SetViewMatrix requested view matrix, position and scale. Position
is useless at it could be found in the inverse view matrix which is computed
anyway. Scale is not needed for users of this function not using any scale
for example shadow and texture renderers.

In consideration the user could call :
SetViewMatrix(viewmat)
SetViewMatrix(viewmat, scale)

The second function test if the scale is negative and negate the rows of the
view matrix.